requirements.txt
anthropic>=0.40.0
boto3>=1.34.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
kubernetes>=29.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Packaging
pyinstaller>=6.0.0
